Modifications to the ldap scripts to bring them back in line with what Kelly is doing with the evaluation process.
