+++ Pipfile
@@ -38,7 +38,7 @@ recommonmark = "*"
 requests = "*"
 sentry-sdk = {extras = ["flask"],version = "==0.14.4"}
 sphinx = "*"
-spiffworkflow = {git = "https://github.com/sartography/SpiffWorkflow.git",ref = "master"}
+spiffworkflow = {git = "https://github.com/sartography/SpiffWorkflow.git",ref = "cr-connect-106-augment-eval"}
 #spiffworkflow = {editable = true,path="/home/kelly/sartography/SpiffWorkflow/"}
 swagger-ui-bundle = "*"
 webtest = "*"
